// @types/StocksCalculatedDataType.d.ts
export type StocksCalculatedDataType = {
  ticker: string;
  price: number;
  dy: number;
  pl: number;
  lpa: number;
  vpa: number;
  dpa: number;
  risk: number;
  discount_margin: number;
  payout: number;
  growthDividend: string;
  roe: number;
  cagrProfit: number;
  damodaramGrowth: number;
  growthAverage: number;
  bazinDiscount: number;
  bazinFairPrice: number;
  bazinCeelingPrice: number;
  grahamDiscount: number;
  grahamFairPrice: number;
  grahamCeelingPrice: number;
  gordonDiscount: number;
  gordonFairPrice: number;
  gordonCeelingPrice: number;
  d1: number;
  peg: number;
  psr: number;
};